Pin setuptools in build requirements (#136)
We ran into an issue in composer where a setuptools release broke pip install from source. This is because pip uses an isolated build environment, which would always install the latest setuptools. In this PR we pin the setuptools version for building the package so that we have control over when to upgrade. Tested pip install -e '.[all]' locally.

See https://github.com/mosaicml/composer/pull/1926