Fix import paths in `simulation` module
Prior to 93af28, the `simulation` directory was added to `PYTHONPATH` by
appending to `sys.path` at the start of each file, before submodules of
`simulation` such as `core` and `interfaces` were imported.

In 93af28, the calls to `sys.path.append` were moved below the import
block and so relative imports of `core` and `interfaces` no longer
resolve and the simulator fails to start with errors such as:

```
  File "/home/scott/projects/streaming/.venv/lib64/python3.12/site-packages/streamlit/runtime/scriptrunner/exec_code.py", line 88, in exec_func_with_error_handling
    result = func()
             ^^^^^^
  File "/home/scott/projects/streaming/.venv/lib64/python3.12/site-packages/streamlit/runtime/scriptrunner/script_runner.py", line 579, in code_to_exec
    exec(code, module.__dict__)
  File "/home/scott/projects/streaming/simulation/interfaces/sim_ui.py", line 17, in <module>
    from core.create_index import create_stream_index
ModuleNotFoundError: No module named 'core'
```

This commit prefixes the submodule imports with `simulation.` so the
imports can be resolved, and without needing to manipulate `sys.path`.
This works even when running the simulator outside the source
repository, as `setuptools.find_packages` is used in `setup.py` to find
the modules to package, and this finds the `simulation` module.